// Develop/SourceCode/A_Frontend/FASS.Web.Vue/src/views/task/tasktemplaterule/utils/types.ts
interface FormItemProps {
  code: string
  name: string
  type: string
  carTypeId: string
  priority: number
  isLoop: boolean,
  timeout: number
  isEnable: boolean,
  remark: string,
  description: string,
  value: string,
  taskTemplateId?: string
}
interface FormProps {
  formInline: FormItemProps;
}

export type { FormItemProps, FormProps };